Clean up object tests to only run once
